+++ uk_address_matcher/cleaning/cleaning_pipelines.py
@@ -22,7 +22,7 @@ from uk_address_matcher.cleaning.cleaning_steps import (
     trim_whitespace_address_and_postcode,
     upper_case_address_and_postcode,
     use_first_unusual_token_if_no_numeric_token,
-    derive_address_token_frequency_table,
+    get_token_frequeny_table,
 )
 from uk_address_matcher.cleaning.run_pipeline import run_pipeline
 
@@ -146,34 +146,60 @@ def clean_data_using_precomputed_rel_tok_freq(
     return con.table(materialised_cleaned_table_name)
 
 
-def create_address_token_frequency_table(
-    address_table: DuckDBPyRelation,
+def get_numeric_term_frequencies_from_address_table(
+    df_address_table: DuckDBPyRelation,
     con: DuckDBPyConnection,
 ) -> DuckDBPyRelation:
-    cleaning_queue = QUEUE_PRE_TF + [
-        derive_address_token_frequency_table,
+    cleaning_queue = [
+        trim_whitespace_address_and_postcode,
+        upper_case_address_and_postcode,
+        clean_address_string_first_pass,
+        parse_out_numbers,
     ]
 
-    # If the following create temp table is not included
-    # and `address_table` is created from like
-    # select * from read_parquet() order by random()
-    # the rest does not work
-
-    uid = _generate_random_identifier()
-    con.register("__address_table_in", address_table)
-
-    materialised_table_name = f"__address_table_{uid}"
-    sql = f"""
-    create or replace temporary table {materialised_table_name} as
-    select * from __address_table_in
-    """
-    con.execute(sql)
-    input_table = con.table(materialised_table_name)
-
-    res = run_pipeline(
-        input_table,
-        con=con,
+    numeric_tokens_df = run_pipeline(
+        df_address_table,
         cleaning_queue=cleaning_queue,
+        print_intermediate=False,
+        con=con,
     )
+    con.register("numeric_tokens_df", numeric_tokens_df)
 
-    return res
+    # For splink, table needs to be in the format
+    # numeric_token__2, tf_numeric_token_2_l
+    sql = """
+    with unnested as
+    (select unnest(numeric_tokens)  as numeric_token
+    from numeric_tokens_df)
+
+    select
+    numeric_token,
+    count(*)/(select count(*) from unnested) as tf_numeric_token
+    from  unnested
+    group by numeric_token
+    order by 2 desc
+
+    """
+    return con.sql(sql)
+
+
+def get_address_token_frequencies_from_address_table(
+    df_address_table: DuckDBPyRelation, con: DuckDBPyConnection
+):
+    cleaning_queue = [
+        trim_whitespace_address_and_postcode,
+        upper_case_address_and_postcode,
+        clean_address_string_first_pass,
+        parse_out_numbers,
+        clean_address_string_second_pass,
+        split_numeric_tokens_to_cols,
+        tokenise_address_without_numbers,
+        get_token_frequeny_table,
+    ]
+
+    return run_pipeline(
+        df_address_table,
+        cleaning_queue=cleaning_queue,
+        print_intermediate=False,
+        con=con,
+    )
